Create saneley named constant subs for error codes
The previous code would create constant subs named
 "address family not supported"
instead of
 EAFNOSUPPORT

and since the description of more than one error code was
  "address family not supported"
we'd get a warning about redifining subs when using the module.

Maybe libuv-0.1 was different in this regard or maybe the bug was
invisible since there was a 1-1 mapping between error codes and error
descriptions.